refactor: migrationとDBカラム定義を整理

+++ db/migrate/20260405220834_add_receipt_detail_fields_to_receipts.rb
@@ -2,8 +2,8 @@ class AddReceiptDetailFieldsToReceipts < ActiveRecord::Migration[8.1]
   def change
     add_column :receipts, :store_address, :text
     add_column :receipts, :store_phone_number, :string
-    add_column :receipts, :subtotal_amount, :integer
-    add_column :receipts, :tax_amount, :integer
+    add_column :receipts, :subtotal_amount, :bigint
+    add_column :receipts, :tax_amount, :bigint
     add_column :receipts, :tax_rate, :integer
   end
 end
