+++ db/migrate/20260426014202_add_discount_amount_to_receipt_items.rb
@@ -1,5 +1,5 @@
 class AddDiscountAmountToReceiptItems < ActiveRecord::Migration[8.1]
   def change
-    add_column :receipt_items, :discount_amount, :integer
+    add_column :receipt_items, :discount_amount, :bigint
   end
 end
